refactor: 탭 데이터 세팅 useTab 훅 분리 및 MainTab → DefaultTab 파일명 수정 (#29)
- 탭 데이터 및 기본 탭 상테 세팅 로직을 커스텀 훅으로 분리
